- Fix a Handshake RTMFP issue regression of novembre 2014 (back to 100% RTMFP handshake success) - Change the FlashStream implementation to match the RTMP Stream specification (streams are listed by session and not by server instance) - Fix a metadata/custom data issue on publication transfer - Add a magic empty audio packet to sync video/audio tracks when stream is pure video - Flash Stream Begin/End signals implemented as RTMP spec. - Supports of NetStream:seek, NetStream:pause and NetStream:resume methods - Publication/Listener frameworks more robust and with a better timestamp evaluation - Fix a "application termination signal" issue on linux target. - Minor change: add a Publication::lastTime() and Entities::empty() function. - Small fixes related to compiler Clang warnings. - API documentation fix and updates

+++ MonaBase/include/Mona/Logs.h
@@ -83,7 +83,7 @@ private:
 	static DumpMode		_DumpMode;
 	static UInt8		_Level;
 	static Logger		_DefaultLogger;
-	static Int16		_DumpLimit; // -1 means no limit
+	static Int32		_DumpLimit; // -1 means no limit
 };
 
 #undef ERROR
